Update Dade.py
Adds conditionals and headless Chrome webdriver. Still need to do the parser.

+++ Dade.py
@@ -9,19 +9,23 @@ import re
 import time
 import csv
 import urllib, json
-
-
-driver = webdriver.Firefox()
+import glob
+import os.path
+options = webdriver.ChromeOptions()
+options.add_argument('headless')
+driver = webdriver.Chrome(chrome_options=options)
 driver.get("http://results.enr.clarityelections.com/FL/Dade/76635")
 time.sleep(10)
 pageSection = driver.find_element_by_class_name("list-download")
 resultLink = pageSection.find_element_by_tag_name("a")
 url = resultLink.get_attribute('href')
 importantNumber = url.split("/")[-3]
-# if important number in json is the same as what's in our folder, break
-#else, open new link
-newUrl = ("http://results.enr.clarityelections.com/FL/Dade/76635/" + str(importantNumber) + "/json/sum.json")
-response = urllib.urlopen(newUrl)
-data = json.loads(response.read())
-with open('Dade%s.json' %importantNumber, 'w') as f:
-    json.dump(data, f)
+#checks to see if the file is already in the directory
+if os.path.isfile('snapshots/Dade%s.json' %importantNumber):
+    print("No new results")
+else:
+    newUrl = ("http://results.enr.clarityelections.com/FL/Dade/76635/" + str(importantNumber) + "/json/sum.json")
+    response = urllib.urlopen(newUrl)
+    data = json.loads(response.read())
+    with open('snapshots/Dade%s.json' %importantNumber, 'w') as f:
+        json.dump(data, f)
